pyproject.toml
[project]
name = "memory-layer"
version = "2.0.0"
description = "高性能AI记忆层服务"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.0",
    "mem0ai[graph]>=0.1.113",
    "chromadb>=1.0.15",
    "fastapi>=0.115.14",
    "uvicorn[standard]>=0.35.0",
    "python-multipart>=0.0.20",
    "aiohttp>=3.12.13",
    "redis>=5.0.0", # 可选：用于分布式缓存
    "prometheus-client>=0.20.0", # 可选：监控指标
    "slowapi>=0.1.9",
    "httpx[socks]>=0.28.1",
    "langchain-openai>=0.3.31",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = ["api", "config", "core", "utils"]
